compose: Add tooltip to provide better feedback when message is invalid.
When a message can not be sent, we serve a error compose banner to the users
to inform them about the error. However, it would provide more consistency and
clarity if we give the user some form of feedback before they try to send a message.

Add a tooltip to the send button which will validate the compose message and
communicate to the user if there are missing requirements or invalid inputs.

Fixes part of #32115.